Split utils/prepare to make TextModels work on Vite project
