+++ src/multimodal/img2text.ts
@@ -7,7 +7,8 @@ import { GenerationConfig } from "../utils/generationConfig";
 import { loadTokenizer } from "../text/tokenizer";
 import { BaseMultimodalModel } from "./base";
 import { ImageProcessingResult } from "../image/interfaces";
-import { prepareImagesTensor, prepareTextTensors } from "../utils/prepare";
+import { prepareTextTensors } from "../utils/prepare-text";
+import { prepareImagesTensor } from "../utils/prepare-image";
 import PreprocessorConfig from "../image/preprocessorConfig";
 import Preprocessor from "../image/preprocessor";
 import * as ort from "onnxruntime-web";
